Change canceled appl import (last by date)

+++ federal_empl_program/imports.py
@@ -178,9 +178,13 @@ def load_application(sheet_dict, row, applicant):
             application = update_application(sheet_dict, row, applicant, application_date)
             return [applications[0], "Updated"]
         elif len(dublicate_applications) > 0:
-            dublicate_applications[0].delete()
-            application = add_application(sheet_dict, row, applicant)
-            return [application, 'Added']
+            if dublicate_applications[0].creation_date < application_date:
+                dublicate_applications[0].delete()
+                application = add_application(sheet_dict, row, applicant)
+                return [application, 'Added']
+            else:
+                application = update_application(sheet_dict, row, applicant, application_date)
+                return [applications[0], "Updated"]
         else:
             application = add_application(sheet_dict, row, applicant)
             return [application, 'Added']
